generalizing the backend for multiple devices and types

+++ include/autograd/autograd_utils.h
@@ -5,8 +5,8 @@
 
 // Struct to pass min_val and max_val to clip_grad_op
 typedef struct {
-  float min_val;
-  float max_val;
+  double min_val;
+  double max_val;
 } ClipExtras;
 
 int get_reduced_dim(int *in_shape, int *out_shape, int in_ndim, int out_ndim);
